# lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Repody",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/runs/{run_id}/status": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Run Status",
        "description": "Lightweight poll \u2014 status + progress only (no full audit payload).",
        "operationId": "get_run_status_v1_runs__run_id__status_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{run_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Run",
        "operationId": "get_run_v1_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "full",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "If false, same as /status (lightweight).",
              "default": true,
              "title": "Full"
            },
            "description": "If false, same as /status (lightweight)."
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunPollResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{run_id}/events": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Stream Run Events",
        "description": "Server-Sent Events stream for live run progress (Redis pub/sub).",
        "operationId": "stream_run_events_v1_runs__run_id__events_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workflows/{workflow_id}/runs/json": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Create Run Json",
        "description": "Create a run with files already uploaded to storage (presigned PUT flow).",
        "operationId": "create_run_json_v1_workflows__workflow_id__runs_json_post",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mode"
            }
          },
          {
            "name": "inline",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Process synchronously (test/dev only).",
              "default": false,
              "title": "Inline"
            },
            "description": "Process synchronously (test/dev only)."
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunJsonBody"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/RunCreatedResponse"
                    },
                    {
                      "$ref": "#/components/schemas/RunPollResponse"
                    }
                  ],
                  "title": "Response Create Run Json V1 Workflows  Workflow Id  Runs Json Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workflows/{workflow_id}/runs": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Create Run",
        "operationId": "create_run_v1_workflows__workflow_id__runs_post",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mode"
            }
          },
          {
            "name": "inline",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Process synchronously (test/dev only).",
              "default": false,
              "title": "Inline"
            },
            "description": "Process synchronously (test/dev only)."
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_run_v1_workflows__workflow_id__runs_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/RunCreatedResponse"
                    },
                    {
                      "$ref": "#/components/schemas/RunPollResponse"
                    }
                  ],
                  "title": "Response Create Run V1 Workflows  Workflow Id  Runs Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workflows/{workflow_id}/run": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Create Run Legacy",
        "operationId": "create_run_legacy_v1_workflows__workflow_id__run_post",
        "deprecated": true,
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_run_legacy_v1_workflows__workflow_id__run_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunCreatedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workflows/{workflow_id}/test-run": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Test Run",
        "description": "Deprecated \u2014 use POST /runs/json?mode=test&inline=true.",
        "operationId": "test_run_v1_workflows__workflow_id__test_run_post",
        "deprecated": true,
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunSnapshotBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_create_run_legacy_v1_workflows__workflow_id__run_post": {
        "properties": {
          "files": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "contentMediaType": "application/octet-stream"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Files"
          },
          "document_ids": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Ids"
          }
        },
        "type": "object",
        "title": "Body_create_run_legacy_v1_workflows__workflow_id__run_post"
      },
      "Body_create_run_v1_workflows__workflow_id__runs_post": {
        "properties": {
          "files": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "contentMediaType": "application/octet-stream"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Files"
          },
          "document_ids": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Ids"
          },
          "payload": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload"
          }
        },
        "type": "object",
        "title": "Body_create_run_v1_workflows__workflow_id__runs_post"
      },
...
      "CreateRunJsonBody": {
        "properties": {
          "snapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunSnapshotBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "payload": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunSnapshotBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "fileBindings": {
            "items": {
              "$ref": "#/components/schemas/StoredFileBinding"
            },
            "type": "array",
            "title": "Filebindings",
            "default": []
          }
        },
        "type": "object",
        "title": "CreateRunJsonBody"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RunCreatedResponse": {
        "properties": {
          "runId": {
            "type": "string",
            "title": "Runid"
          },
          "jobId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jobid"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "queued"
          }
        },
        "type": "object",
        "required": [
          "runId"
        ],
        "title": "RunCreatedResponse"
      },
...
      "RunPollResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunProgressSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunAuditDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "RunPollResponse"
      },
...
      "RunSnapshotBody": {
        "properties": {
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentDefSchema"
            },
            "type": "array",
            "title": "Documents",
            "default": []
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/WorkflowRuleSchema"
            },
            "type": "array",
            "title": "Rules",
            "default": []
          },
          "workflow_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Name"
          }
        },
        "type": "object",
        "title": "RunSnapshotBody",
        "description": "Ephemeral workflow config for this run only \u2014 does not mutate the saved workflow."
      },
...
      "StoredFileBinding": {
        "properties": {
          "documentId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Documentid"
          },
          "storageKey": {
            "type": "string",
            "title": "Storagekey"
          },
          "mimeType": {
            "type": "string",
            "title": "Mimetype"
          },
          "fileName": {
            "type": "string",
            "title": "Filename"
          }
        },
        "type": "object",
        "required": [
          "storageKey",
          "mimeType",
          "fileName"
        ],
        "title": "StoredFileBinding"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}